processor/apps: fpm apps processor init
